Add ungrouped_evaluator option to attribute grouped source parser
Add an optional ungrouped_evaluator parameter that allows handling code
blocks that don't have the grouping attribute. When provided, these
blocks are evaluated with the ungrouped evaluator instead of being
ignored.

This is useful when you want to process all code blocks in a document
but still group certain blocks together.